% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageOxygen.R
\name{Biweight.kernel}
\alias{Biweight.kernel}
\title{Biweight kernel function}
\usage{
Biweight.kernel(x)
}
\arguments{
\item{x}{a vector.}
}
\description{
Biweight kernel function.
}
\details{
Biweight kernel:
\deqn{
  K(x) = 15/16 ( 1 - x^2 )^2  (abs(x)<=1)
}
We recommend a critical value of 7 for this kernel function.
}
\examples{
plot(function(x) Biweight.kernel(x),-2, 2,
main = " Biweight kernel ")

}
