% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageOxygen.R
\name{Epa.kernel}
\alias{Epa.kernel}
\title{Epanechnikov kernel function}
\usage{
Epa.kernel(x)
}
\arguments{
\item{x}{vector.}
}
\description{
Epanechnikov kernel function.
}
\details{
Epanechnikov kernel:
\deqn{
  K(x) = 3/4 ( 1 - x^2 ) (abs(x)<=1)
}
We recommend a critical value of 6.1 for this kernel function.
}
\examples{
plot(function(x) Epa.kernel(x), -2, 2, ylab = "Epanechnikov kernel function")

}

