% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageOxygen.R
\name{Gaussian.kernel}
\alias{Gaussian.kernel}
\title{Gaussian kernel function}
\usage{
Gaussian.kernel(x)
}
\arguments{
\item{x}{a vector.}
}
\description{
Gaussian kernel function
}
\details{
Gaussian Kernel with the value of standard deviation equal to 1/3.
\deqn{
  K(x) = (1/{(1/3)*sqrt(2 \pi)}  exp(-(3*x)^2/2)) (abs(x) <= 1)
}
We recommend a critical value of 8.3 for this kernel.
}
\examples{
plot(function(x) Gaussian.kernel(x), -2, 2,
main = " Gaussian kernel")

}

