% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageOxygen.R
\name{probgrid}
\alias{probgrid}
\title{Probability grid}
\usage{
probgrid(p1, p2, length = 50)
}
\arguments{
\item{p1}{the first element of the grid.}

\item{p2}{the last element of the grid.}

\item{length}{the length of the grid.}
}
\value{
A vector of probabilities between \eqn{p1} and \eqn{p2} and length \eqn{length}.
}
\description{
Create a geometric grid of probabilities
}
\details{
Create a geometric grid of length \eqn{length} between \eqn{p1} and \eqn{p2}.The default value of \eqn{length} is \eqn{50}.
}
\examples{
p1 <- 0.01
p2 <- 0.99
length <- 500
pgrid <- probgrid(p1, p2, length)

}

