% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageOxygen.R
\name{Rectangular.kernel}
\alias{Rectangular.kernel}
\title{Rectangular kernel function}
\usage{
Rectangular.kernel(x)
}
\arguments{
\item{x}{a vector.}
}
\description{
Rectangular kernel function
}
\details{
Rectangular kernel function

Rectangular Kernel
\deqn{
  K(x) = 1 (abs(x) <= 1)
}
We recommend a critical value of 10 for this kernel.
}
\examples{
plot(function(x) Rectangular.kernel(x), -2, 2,
main = " Rectangular kernel ")

}
