% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageOxygen.R
\name{Triang.kernel}
\alias{Triang.kernel}
\title{Triangular kernel function}
\usage{
Triang.kernel(x)
}
\arguments{
\item{x}{a vector.}
}
\description{
Triangular kernel function
}
\details{
Triangular Kernel
\deqn{
  K(x) = ( 1 - abs(x) )  (abs(x) <= 1)
}
We recommend a critical value of 6.9 for this kernel.
}
\examples{
plot(function(x) Triang.kernel(x), -2, 2,
main = " Triangular kernel")

}

